--- src/components/NewPost/index.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import './index.css';
-
-const NewPost = () => {
-    return (
-      <div>
-        <form>
-          <input type="string" /><br />
-          <button>投稿</button>
-        </form>
-      </div>
-    )
-}
-
-export default NewPost;
