+++ src/redux/rootReducer.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import posts from './modules/posts';
+
+export default combineReducers({
+    posts
+})
